+++ GUI/nbproject/private/private.properties
@@ -1,2 +1,2 @@
 compile.on.save=true
-user.properties.file=/Users/oscarrodriguez/Library/Application Support/NetBeans/12.6/build.properties
+user.properties.file=C:\\Users\\Gustavo\\AppData\\Roaming\\NetBeans\\12.6\\build.properties
