+++ GUI/src/GUIFrame.java
@@ -210,6 +210,11 @@ public class GUIFrame extends javax.swing.JFrame {
         jButton1.setFont(new java.awt.Font("Courier New", 1, 24)); // NOI18N
         jButton1.setForeground(new java.awt.Color(0, 102, 51));
         jButton1.setText("INICIAR");
+        jButton1.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jButton1ActionPerformed(evt);
+            }
+        });
 
         javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
         getContentPane().setLayout(layout);
@@ -234,6 +239,17 @@ public class GUIFrame extends javax.swing.JFrame {
 
         pack();
     }// </editor-fold>//GEN-END:initComponents
+
+    private void jButton1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton1ActionPerformed
+        // TODO add your handling code here:
+        Buffer buffer = new Buffer();
+        
+        Producer producer = new Producer(buffer);
+        producer.start();
+        
+        Consumer consumer = new Consumer(buffer);
+        consumer.start();
+    }//GEN-LAST:event_jButton1ActionPerformed
 
     /**
      * @param args the command line arguments
